requirements.txt
flask
flask-cors
python-dotenv
google-genai
trimesh
numpy
reportlab